Fixed heal ability bug not consuming potions Fixed fighter bug trying to call an animation that didn't exist Health and bond UI now update when traits are changed

+++ Assets/Scripts/Combat/WeaponConfig.cs
@@ -126,7 +126,7 @@ namespace RPG.Combat
             projectileInstance.SetTarget(target, instigator, calculatedDamage);
         }
 
-        public IEnumerable<int> GetAddativeModifiers(Stat stat)
+        public IEnumerable<float> GetAddativeModifiers(Stat stat)
         {
             if (stat == Stat.Damage)
             {
