+++ Assets/Scripts/Stats/IModifierProvider.cs
@@ -4,7 +4,7 @@ namespace RPG.Stats
 {
     public interface IModifierProvider
     {
-        IEnumerable<int> GetAddativeModifiers(Stat stat);
+        IEnumerable<float> GetAddativeModifiers(Stat stat);
         IEnumerable<float> GetPercentageModifiers(Stat stat);
     }
 }
